requirements.txt
flask
imgurpython
pytesseract
requests
Pillow